+++ spec/lib/danger/ci_sources/ci_source_spec.rb
@@ -18,6 +18,7 @@ RSpec.describe Danger::CI do
           "Danger::CodeBuild",
           "Danger::Codefresh",
           "Danger::Codeship",
+          "Danger::Concourse",
           "Danger::DotCi",
           "Danger::Drone",
           "Danger::GitLabCI",